:recycle:Refactor: refactor data widget 데이터 관련 위젯 모듈화를 위한 리팩토링 Resolves: #18
