+++ coin/src/components/units/main/widget/AvailableWidgets.tsx
@@ -10,6 +10,6 @@ export const availableWidgets = [
     { type: 'dogecoin', name: '도지코인', icon: <>DOGE</>, symbol: 'BINANCE:DOGEUSD', category: 'crypto' },
     { type: 'cardano', name: '에이다', icon: <>ADA</>, symbol: 'BINANCE:ADAUSD', category: 'crypto' },
     // 데이터 시각화 위젯
-    //{ type: 'binance', name: '바이낸스 선물 비트코인 롱/숏 포지션 비율', icon: <></>, symbol: undefined, category: 'data' },
-    { type: 'binance', name: '공포 / 탐욕 지수', icon: <></>, symbol: undefined, category: 'data' },
+    { type: 'longshortratio', name: '바이낸스 선물 비트코인 롱/숏 포지션 비율', icon: <></>, symbol: undefined, category: 'data' },
+    { type: 'feargreedindex', name: '공포 / 탐욕 지수', icon: <></>, symbol: undefined, category: 'data' },
 ];
